Refactor download_blinks into utils.py
