+++ tests/test_core.py
@@ -1,8 +1,4 @@
-import pathlib
-
 import cloudscraper
-
-from dailyblink.__main__ import download_blinks
 
 from dailyblink.core import (
     get_daily_blink_info,
@@ -11,7 +7,6 @@ from dailyblink.core import (
     save_audio_content,
     save_book_text,
 )
-from dailyblink.settings import AVAILABLE_LANGUAGES
 
 scraper = cloudscraper.create_scraper()
 
@@ -78,7 +73,3 @@ def test_save_book_text():
     blink_url = blink_info["url"]
     chapters = request_blinkist_book_text(scraper, blink_url)["chapters"]
     save_book_text(blink_info, chapters, file_path="test_output/daily_blink.md")
-
-
-def test_download_blinks():
-    download_blinks(scraper, AVAILABLE_LANGUAGES, pathlib.Path(".") / "blinks")
